+++ phosphor-power-sequencer/src/standard_device.hpp
@@ -65,7 +65,7 @@ class StandardDevice : public BasicDevice
      *                          signal from this device
      * @param rails voltage rails that are enabled and monitored by this device
      */
-    explicit StandardDevice(const std::string& name, uint8_t bus,
+    explicit StandardDevice(const std::string& name, uint16_t bus,
                             uint16_t address,
                             const std::string& powerControlGPIOName,
                             const std::string& powerGoodGPIOName,
