+++ phosphor-power-sequencer/test/ucd90160_device_tests.cpp
@@ -59,7 +59,7 @@ static std::unique_ptr<Rail> createRail(const std::string& name,
 
 TEST(UCD90160DeviceTests, Constructor)
 {
-    uint8_t bus{3};
+    uint16_t bus{3};
     uint16_t address{0x72};
     std::string powerControlGPIOName{"power-chassis-control"};
     std::string powerGoodGPIOName{"power-chassis-good"};
@@ -149,7 +149,7 @@ TEST(UCD90160DeviceTests, StoreGPIOValues)
                 "Rail VDD pgood GPIO line offset 2 has inactive value 0"))
             .Times(1);
 
-        uint8_t bus{3};
+        uint16_t bus{3};
         uint16_t address{0x72};
         std::string powerControlGPIOName{"power-chassis-control"};
         std::string powerGoodGPIOName{"power-chassis-good"};
@@ -250,7 +250,7 @@ TEST(UCD90160DeviceTests, StoreGPIOValues)
                 "Rail VDD pgood GPIO line offset 2 has inactive value 0"))
             .Times(1);
 
-        uint8_t bus{3};
+        uint16_t bus{3};
         uint16_t address{0x72};
         std::string powerControlGPIOName{"power-chassis-control"};
         std::string powerGoodGPIOName{"power-chassis-good"};
